requirements.txt
attr==0.3.2
attrs==18.1.0
brotli==1.0.9
brotlicffi==1.0.9.2
brotlipy==0.7.0
configparser==3.5.0
cryptography==36.0.2
Cython==0.29.28
dl==0.1.0
docutils==0.18.1
HTMLParser==0.0.2
importlib_metadata==4.8.1
ipython==8.6.0
ipywidgets==7.7.0
Jinja2==3.1.0
jnius==1.1.0
keyring==23.4.0
lockfile==0.12.2
lxml==4.8.0
matplotlib==3.5.2
mock==4.0.3
numpy==1.21.5
opencv_python==4.5.5.64
ordereddict==1.1
Pillow==9.3.0
protobuf==4.21.9
pyOpenSSL==22.1.0
pyvista==0.35.1
pywin32==303
railroad==0.5.0
simplejson==3.18.0
Sphinx==4.4.0
toml==0.10.2
tornado==6.2
unicodedata2==14.0.0
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
zipp==3.7.0
